move: @type/BlogPostType -> feature blog folder
