+++ src/features/blog/types/index.ts
@@ -1,0 +1,1 @@
+export * as BlogPostTypes from './BlogPost';
