+++ src/types/index.ts
@@ -1,3 +1,2 @@
 export * as ProductTypes from './Product';
 export * as CartTypes from './Cart';
-export * as BlogPostTypes from './BlogPost';
